Move the function to operator directory from bundle.
