+++ operator-pipeline-images/operatorcert/schemas/ci-schema.json
@@ -8,8 +8,28 @@
             "description": "Config set for FBC-enabled operator bundles",
             "type": "object",
             "properties": {
+                "updateGraph": {
+                    "description": "Name of the update strategy for the operator",
+                    "type": "string"
+                },
+                "reviewers": {
+                    "description": "List of authorized GitHub usernames",
+                    "type": "array",
+                    "items": {
+                        "type": "string",
+                        "minItems": 1
+                    }
+                },
+                "merge": {
+                    "description": "Flag that determines whether the PR is set to auto-merge",
+                    "type": "boolean"
+                },
+                "cert_project_id": {
+                    "description": "Certification Project ID linked with the operator",
+                    "type": "string"
+                },
                 "enabled": {
-                    "description": "Whether the operator already have FBC support or not",
+                    "description": "Flag that determines whether the operator uses FBC delivery method",
                     "type": "boolean"
                 },
                 "catalog_mapping": {
@@ -19,7 +39,7 @@
                         "type": "object",
                         "properties": {
                             "template_name": {
-                                "description": "Name of the catalog template file (required)",
+                                "description": "Name of the catalog template file",
                                 "type": "string"
                             },
                             "catalog_names": {
@@ -32,7 +52,8 @@
                             },
                             "type": {
                                 "description": "Type of template schema",
-                                "type": "string"
+                                "type": "string",
+                                "enum": ["olm.template.basic", "olm.semver"]
                             }
                         },
                         "required": [ "template_name", "type", "catalog_names" ]
@@ -43,6 +64,5 @@
             "minItems": 1,
             "required": [ "catalog_mapping" ]
         }
-    },
-    "required": [ "fbc" ]
+    }
 }
