+++ operator-pipeline-images/operatorcert/static_tests/common/bundle.py
@@ -167,26 +167,6 @@ def check_bundle_release_config(bundle: Bundle) -> Iterator[CheckResult]:
             )
 
 
-def check_validate_schema_ci_config(
-    bundle: Bundle,
-) -> Iterator[CheckResult]:
-    """
-    Validate the ci.yaml against the json schema
-    """
-    path_me = os.path.dirname(os.path.abspath(__file__))
-    path_schema = os.path.join(path_me, "../../schemas/ci-schema.json")
-    with open(path_schema, "r", encoding="utf-8") as file_schema:
-        dict_schema = json.load(file_schema)
-    # validate the ci.yaml against the json schema
-    # use iter_errors() to collect and return all validation errors
-    validator = Draft202012Validator(dict_schema)
-    for ve in sorted(validator.iter_errors(bundle.operator.config), key=str):
-        yield Fail(
-            "Bundle's 'ci.yaml' contains invalid data "
-            f"which does not comply with the schema: {ve.message}"
-        )
-
-
 def check_validate_schema_bundle_release_config(
     bundle: Bundle,
 ) -> Iterator[CheckResult]:
