+++ operator-pipeline-images/tests/static_tests/common/test_bundle.py
@@ -7,7 +7,6 @@ from operatorcert.static_tests.common.bundle import (
     check_operator_name,
     check_bundle_release_config,
     check_validate_schema_bundle_release_config,
-    check_validate_schema_ci_config,
 )
 from tests.utils import bundle_files, create_files
 
@@ -651,177 +650,4 @@ def test_check_validate_schema_bundle_release_config(
     assert {
         (x.__class__, x.reason)
         for x in check_validate_schema_bundle_release_config(bundle)
-    } == expected_results
-
-
-@pytest.mark.parametrize(
-    "files, bundle_to_check, expected_results",
-    [
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {"operators/hello/ci.yaml": {"key": "value"}},
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'fbc' is a required property",
-                ),
-            },
-            id="Testing validation without fbc object",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {"operators/hello/ci.yaml": {"fbc": "hello"}},
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'hello' is not of type 'object'",
-                ),
-            },
-            id="Testing the validation for invalid fbc field type",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {"operators/hello/ci.yaml": {"fbc": {"enabled": True}}},
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'catalog_mapping' is a required property",
-                ),
-            },
-            id="Testing the validation for missing catalog_mapping field",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {
-                    "operators/hello/ci.yaml": {
-                        "fbc": {"enabled": True, "catalog_mapping": "hello"}
-                    }
-                },
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'hello' is not of type 'array'",
-                ),
-            },
-            id="Testing the validation for invalid catalog_mapping array data",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {
-                    "operators/hello/ci.yaml": {
-                        "fbc": {"enabled": True, "catalog_mapping": [{"hello": ""}]}
-                    }
-                },
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'template_name' is a required property",
-                ),
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'type' is a required property",
-                ),
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'catalog_names' is a required property",
-                ),
-            },
-            id="Testing with missing required fields in catalog_mapping array",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {
-                    "operators/hello/ci.yaml": {
-                        "fbc": {
-                            "enabled": True,
-                            "catalog_mapping": [{"type": "demo.type"}],
-                        }
-                    }
-                },
-            ],
-            ("hello", "0.0.1"),
-            {
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'template_name' is a required property",
-                ),
-                (
-                    Fail,
-                    "Bundle's 'ci.yaml' contains invalid data "
-                    "which does not comply with the schema: "
-                    "'catalog_names' is a required property",
-                ),
-            },
-            id="Testing with missing required fields in catalog_mapping array",
-        ),
-        pytest.param(
-            [
-                bundle_files("hello", "0.0.1"),
-                {
-                    "operators/hello/ci.yaml": {
-                        "fbc": {
-                            "enabled": True,
-                            "catalog_mapping": [
-                                {
-                                    "template_name": "foo",
-                                    "catalog_names": ["1.0", "2.0"],
-                                    "type": "demo.type",
-                                }
-                            ],
-                        }
-                    }
-                },
-            ],
-            ("hello", "0.0.1"),
-            set(),
-            id="Testing with all valid fields with required data for ci.yaml schema.",
-        ),
-    ],
-    indirect=False,
-)
-def test_check_validate_schema_ci_config(
-    tmp_path: Path,
-    files: list[dict[str, Any]],
-    bundle_to_check: tuple[str, str],
-    expected_results: set[tuple[type, str]],
-) -> None:
-    create_files(tmp_path, *files)
-    repo = Repo(tmp_path)
-    operator_name, bundle_version = bundle_to_check
-    operator = repo.operator(operator_name)
-    bundle = operator.bundle(bundle_version)
-    assert {
-        (x.__class__, x.reason) for x in check_validate_schema_ci_config(bundle)
     } == expected_results
